Test: Fix EVO_PACKAGE_DIRECTORY for the test suite
It's awkward having to do this, but normally user scripts will be run from the bundle's root and this shouldn't be needed. If they do need to deviate from this convention, I suppose they'd also have to adjust the package folder so that it points to the right path relative to their entry point.

For now, however, only main.lua is supported so this is admittedly a bit of  hack (but a harmless one).