Refactor: Move the entry point of the test runner
This works around an issue with relative imports from a subdirectory, and is consistent with the top-level script commands (start = main.lua, test = test.lua) all being stored in the root directory.

Not sure if this convention should be adopted everywhere, but until I find a better design I guess it'll do.