Core: Add an extension endpoint for the uv library
The default libuv error codes aren't always very comprehensive, so displaying the full error message seems like a good idea. The luv bindings already do this in many cases, but it doesn't seem to be working for callbacks, only library/API calls, which is somewhat unfortunate.

The alternative would be exposing uv_strerror on the luv side, but that's more involved and the function may leak memory (according to the libuv docs). Maybe later? For now, this will do.