Core: Fix import errors when loading from the bundle VFS
The import extension (from evo-luvi) currently doesn't support nested imports for bundle paths. Reworking that will take some effort, so in the meantime this change will restore functionality. It's obviously a hack.

Yes, this is bad, but the problem needs to be fixed properly at the source, so there's not much I can do here until the module loading mechanism works properly when loading from the bundle VFS.